Break public API for 2.0: offsets, lifecycle, positions, content-only Buffer
Four coordinated breaking changes, each functionally gated (full verify,
334+16 tests) and performance-gated on the agogo Docker rig
(stable_10k_moderate, ratio gate 1.10, identical match counts):

- Match callbacks now use half-open [start, end offsets, matching
  String.substring and Buffer.getString. Engines stay inclusive
  internally; the single conversion point is RegexpImpl.performActions.
- Matcher extends AutoCloseable. shutdown() is replaced by an
  idempotent no-throw close(); partitioned matchers force-terminate
  their pool after a grace period and restore the interrupt flag.
- Buffer positions are long across the public API (getCurrentPos,
  getString, Action offsets); engine internals keep int positions and
  widen at the callback boundary.
- Buffer is now pure content: hasCharAt(long), charAt(long), and
  getString(long, long) with a default implementation. No total-length
  method, so bounded-window buffers over streaming input remain
  expressible. clone(), Comparable, the cursor methods, and
  LookaheadBuffer are gone; engines keep local long cursors, partitions
  share one buffer instance, and RegexStringBuffer drops its
  per-character synchronization and Character boxing.

Perf receipts: parity on 1MB and 10MB (worst like-for-like ratio 1.044).
An apparent 2x speedup in the first runs was traced to a first-10MB-cell
warm-up artifact on the benchmark host and debunked with an
order-swapped run. Gate logs preserved on agogo under
rmatch-agogo-v2-breaking-gate-20260709_212913.

EOF
)

+++ rmatch/src/main/java/no/rmz/rmatch/Action.java
@@ -17,13 +17,14 @@ package no.rmz.rmatch;
  * Callback invoked when a registered expression matches an input buffer.
  *
  * <p>Implement this interface to collect matches, update counters, emit records, or otherwise make
- * matches visible to application code. The callback receives the input buffer and inclusive
- * start/end offsets. To recover the matched text, call {@code buffer.getString(start, end + 1)}.
+ * matches visible to application code. The callback receives the input buffer and half-open {@code
+ * [start, end)} offsets, the same convention as {@link String#substring(int, int)}. To recover the
+ * matched text, call {@code buffer.getString(start, end)}.
  *
  * <p><b>Thread-safety contract:</b> actions may be invoked concurrently from multiple engine worker
  * threads. In particular, {@code RMatch.newMatcher()} returns a partitioned matcher on multi-core
  * systems. An action instance, especially one shared between several expressions, must therefore be
- * prepared for concurrent {@link #performMatch(Buffer, int, int)} invocations. Use {@code
+ * prepared for concurrent {@link #performMatch(Buffer, long, long)} invocations. Use {@code
  * java.util.concurrent} types such as {@code LongAdder} or {@code AtomicLong}, a synchronized
  * block, or a concurrent collection for mutable state. A plain {@code int++} counter in an action
  * is a lost-update bug waiting to happen.
@@ -33,16 +34,20 @@ public interface Action {
   /**
    * Handle a single match.
    *
-   * <p>The {@code start} and {@code end} offsets are inclusive. This is deliberately different from
-   * {@link Buffer#getString(int, int)}, whose second argument is exclusive like {@link
-   * String#substring(int, int)}.
+   * <p>The match spans the half-open range {@code [start, end)}: {@code start} is inclusive and
+   * {@code end} is exclusive, exactly like {@link String#substring(int, int)} and {@link
+   * Buffer#getString(int, int)}. The matched text is therefore {@code b.getString(start, end)} and
+   * its length is {@code end - start}.
    *
    * <p>This method may be called concurrently from multiple threads; see the thread-safety contract
    * in the class documentation.
    *
+   * <p>Offsets are {@code long} to match {@link Buffer} positions, so custom buffer implementations
+   * larger than the {@code int} range stay expressible.
+   *
    * @param b buffer where the match occurred
    * @param start zero-based inclusive start offset of the match
-   * @param end zero-based inclusive end offset of the match
+   * @param end zero-based exclusive end offset of the match
    */
-  void performMatch(final Buffer b, final int start, final int end);
+  void performMatch(final Buffer b, final long start, final long end);
 }
